requirements.txt
loguru
Flask
flask_cors
PyYAML